refactored url_for's
