+++ flaskblog/templates/post.html
@@ -5,12 +5,12 @@
              src="{{ url_for('static', filename='profile_pics/' + post.author.image_file) }}" alt="">
         <div class="media-body">
             <div class="article-metadata">
-                <a class="mr-2" href="{{ url_for('user_posts', username=post.author.username) }}">{{ post.author.username }}</a>
+                <a class="mr-2" href="{{ url_for('users.user_posts', username=post.author.username) }}">{{ post.author.username }}</a>
                 <small class="text-muted">{{ post.date_posted.strftime('%Y-%m-%d') }}</small>
                 {% if post.author == current_user %}
                     <div>
                         <a class="btn-secondary btn-sm mt-1 mb-1"
-                           href="{{ url_for('post_update', post_id=post.id) }}">Update</a>
+                           href="{{ url_for('posts.post_update', post_id=post.id) }}">Update</a>
                         <button type="button" class="btn btn-danger btn-sm m-1" data-toggle="modal"
                                 data-target="#deleteModal">Delete
                         </button>
@@ -34,7 +34,7 @@
                 </div>
                 <div class="modal-footer">
                     <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
-                    <form action="{{ url_for('post_delete', post_id=post.id) }}" method="POST">
+                    <form action="{{ url_for('posts.post_delete', post_id=post.id) }}" method="POST">
                         <input class="btn btn-danger" type="submit" value="Delete">
                     </form>
                 </div>
